% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_calc_move_llk}
\alias{C_calc_move_llk}
\title{Computes log-likelihood of Jolly-Seber model}
\usage{
C_calc_move_llk(n, J, pr0, pr_capture, tpms, num_cells, inside, dx, dt, sd,
  num_cores, num_states, entry)
}
\arguments{
\item{n}{number of individuals}

\item{J}{total number of occasions}

\item{pr0}{initial distribution over life states}

\item{pr_capture}{output of calc_pr_capture() in JsModel}

\item{tpms}{output of calc_tpms() in JsModel}

\item{num_cores}{number of processor cores to use in parallelisation}

\item{num_states}{2 = CJS model, 3 = JS model}

\item{M}{total number of mesh points}
}
\value{
log-likelihood value
}
\description{
Computes log-likelihood of Jolly-Seber model
}
