% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs_trans_model.R
\docType{data}
\name{CjsTransientModel}
\alias{CjsTransientModel}
\title{Cormack-Jolly-Seber with transience model class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CjsTransientModel
}
\description{
Cormack-Jolly-Seber with transience model: fits model, formats inference, and 
simulates from fitted model. This model inherits all the functions of the 
CjsModel class. Functions that are different or additional are documented here. 
\itemize{
  \item form: a named list of formulae for each parameter (~1 for constant)
  \item scr_data: a ScrData object 
  \item start: a named list of starting values 
  \item num_cores (optional, default = 1): number of processors cores to use 
  in parallelised code 
  \item print: (defualt TRUE) if TRUE then useful output is printed
}

Methods include: 
\itemize{
 \item calc_initial_distribution(): computes initial distribution over life states (unborn, alive, dead)
 \item calc_llk(): compute log-likelihood at current parameter values 
 \item simulate(): simulate ScrData object from fitted model
}
}
\keyword{datasets}
