% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedValues.R
\name{DerivedValues}
\alias{DerivedValues}
\title{Function to post-process MCMC results. Need to extend for model without sex.}
\usage{
DerivedValues(Nm, Nf, gamma, phi, psex, psi, M)
}
\arguments{
\item{Nm}{matrix containing the posteriors for the number of males}

\item{Nf}{matrix containing the posteriors for the number of females}

\item{gamma}{matrix of vector containing the posteriors for gamma(s)}

\item{phi}{matrix of vector containing the posteriors for phi(s)}

\item{psex}{vector containing the posterior for psex}
}
\value{
a list containing realized and expected population growth rates, population sizes,
and sex ratios
}
\description{
This function derives realized and expected population growth rates, population sizes, and sex ratios
from the inputted MCMC chains.
}
\author{
Ben Augustine
}
