% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fn.R
\docType{data}
\name{DetFn}
\alias{DetFn}
\title{Detection function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DetFn
}
\description{
Object represents the encounter function / detection function. 
\itemize{
  \item parameters: a character vector of parameter names
  \item fn: either a character denoting a encounter rate function / detection function or a function
  that takes parameter vector and distance and returns encounter rate or probability 
  \item prob: if fn supplied by user return probability then set this to TRUE, otherwise it is FALSE 
}

Methods include: 
\itemize{
 \item g: returns detection probability at a given distance and parameter values 
 \item h: returns encounter rate at a given distance and parameter values 
 \item plot: plots detection function if prob = TRUE, otherwise encounter function 
 \item print: prints out basic encounter information 
}
}
\keyword{datasets}
