% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoini.R
\name{get_start_values}
\alias{get_start_values}
\title{Get initial values}
\usage{
get_start_values(obj, model = "ScrModel")
}
\arguments{
\item{obj}{ScrData}

\item{model}{model class}
}
\value{
initial values for parameters
}
\description{
Get initial values
}
