% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jsmodel_example}
\alias{jsmodel_example}
\title{Fitted models from JsModel vignette}
\format{A list of two data frames 
\describe{
  \item{mod}{standard SCR model} 
  \item{mod_seas}{SCR model with detector covariate}
  \item{stat}{stationary SCR model to transient data}
  \item{trans}{transient SCR model to transient data}
}}
\usage{
jsmodel_example
}
\description{
Includes all fitted models from vignette.
}
\keyword{datasets}
