% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_trans_model.R
\docType{data}
\name{JsTransientModel}
\alias{JsTransientModel}
\title{Jolly-Seber with transience model class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JsTransientModel
}
\description{
Jolly-Seber model with transience: fits model, formats inference, and 
simulates from fitted model. This model inherits all the functions of the 
JsModel class. Functions that are different or additional are documented here. 
\itemize{
  \item form: a named list of formulae for each parameter (~1 for constant)
  \item scr_data: a ScrData object 
  \item start: a named list of starting values 
  \item num_cores (optional, default = 1): number of processors cores to use 
  in parallelised code 
  \item print (defualt = TRUE): if TRUE then helpful output is printed to the screen
}

Methods include: 
\itemize{
 \item calc_initial_distribution(): computes initial distribution over life states (unborn, alive, dead)
 \item calc_D_llk(): computes the likelihood of the D parameter
 \item calc_pdet(): compute probability of being detected at least once during the survey
 \item calc_llk(): compute log-likelihood at current parameter values 
 \item fit(nsims = 99): fit the model by obtaining the maximum likelihood estimates. Estimates of
       density are obtained from parametric boostrap with nsim resamples. 
 \item simulate(): simulate ScrData object from fitted model
}
}
\keyword{datasets}
