% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.CjsModel}
\alias{logLik.CjsModel}
\title{Compute Log-Likelihood of Cormack-Jolly-Seber Model}
\usage{
\method{logLik}{CjsModel}(object, ...)
}
\arguments{
\item{object}{CjsModel object}

\item{...}{ignored}
}
\value{
log-likelihood
}
\description{
Compute Log-Likelihood of Cormack-Jolly-Seber Model
}
