% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.CjsTransientModel}
\alias{logLik.CjsTransientModel}
\title{Compute Log-Likelihood of Cormack-Jolly-Seber Transient Model}
\usage{
\method{logLik}{CjsTransientModel}(object, ...)
}
\arguments{
\item{object}{CjsModel object}

\item{...}{ignored}
}
\value{
log-likelihood
}
\description{
Compute Log-Likelihood of Cormack-Jolly-Seber Transient Model
}
