% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.JsModel}
\alias{logLik.JsModel}
\title{Compute Log-Likelihood of Jolly-Seber Model}
\usage{
\method{logLik}{JsModel}(object, ...)
}
\arguments{
\item{object}{JsModel object}

\item{...}{ignored}
}
\value{
log-likelihood
}
\description{
Compute Log-Likelihood of Jolly-Seber Model
}
