% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.JsTransientModel}
\alias{logLik.JsTransientModel}
\title{Compute Log-Likelihood of Jolly-Seber Transient Model}
\usage{
\method{logLik}{JsTransientModel}(object, ...)
}
\arguments{
\item{object}{JsModel object}

\item{...}{ignored}
}
\value{
log-likelihood
}
\description{
Compute Log-Likelihood of Jolly-Seber Transient Model
}
