% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.ScrModel}
\alias{logLik.ScrModel}
\title{Compute Log-Likelihood of SCR Model}
\usage{
\method{logLik}{ScrModel}(object, ...)
}
\arguments{
\item{object}{ScrModel object}

\item{...}{ignored}
}
\value{
log-likelihood
}
\description{
Compute Log-Likelihood of SCR Model
}
