% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.ScrTransientModel}
\alias{logLik.ScrTransientModel}
\title{Compute Log-Likelihood of SCR Transient Model}
\usage{
\method{logLik}{ScrTransientModel}(object, ...)
}
\arguments{
\item{object}{ScrTransientModel object}

\item{...}{ignored}
}
\value{
log-likelihood
}
\description{
Compute Log-Likelihood of SCR Transient Model
}
