% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpop_example.R
\docType{data}
\name{openpop_example}
\alias{openpop_example}
\title{Example open population SCR data and fitted models
See Vignette}
\format{An object of class \code{list} of length 5.}
\usage{
openpop_example
}
\description{
Example open population SCR data and fitted models
See Vignette
}
\keyword{datasets}
