% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpop_example_data.R
\docType{data}
\name{openpop_example_data}
\alias{openpop_example_data}
\title{Example open population SCR data for count detectors}
\format{An object of class \code{list} of length 2.}
\usage{
openpop_example_data
}
\description{
Example open population SCR data for count detectors
}
\keyword{datasets}
