% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_data.R
\docType{data}
\name{ScrData}
\alias{ScrData}
\title{Spatial capture-recapture data class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScrData
}
\description{
Encapsulates data from spatial capture-recapture survey into a single 
object. Object can be created using $new with arguments: 
\itemize{
  \item capthist: capture history object from secr package with trap information included
  \item mesh: mask object from secr package
  \item time: optional vector of numeric time each occasion took place at 
    (used for irregularly spaced occasions). Default is vector of ones.
  \item cov: a list of covariate (for detector and occasion)
  \item cov_type: type of covariate (j = detector cov, k = occasion cov, jk = detector+occasion,)
       
}

Methods include: 
\itemize{
 \item capthist(): return capture history object 
 \item traps(): return traps object 
 \item mesh(): return mesh object
 \item time(): return time vector 
 \item covs(j, k): return covariate values at detector j 
       on occasion k  
 \item n(): return number of individuals seen over the entire survey
 \item n_occasions(): return number of capture occasions in the survey
 \item n_traps(): return number of traps used at some time in the survey
 \item n_meshpts(): return number of mesh points 
 \item area(): return total area of the mesh 
 \item add_covariate(covariate_name, covariate_dataframe, covariate_type): add covariate to object
 \item distance(): return matrix with (i,j)th entry the distance from trap i to mesh point j
}
}
\keyword{datasets}
