% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrmodel_example}
\alias{scrmodel_example}
\title{Fitted models from ScrModel vignette}
\format{A list of two data frames 
\describe{
  \item{mod}{standard SCR model} 
  \item{mod_detage}{SCR model with detector covariate}
  \item{stat}{stationary SCR model to transient data}
  \item{trans}{transient SCR model to transient data}
}}
\usage{
scrmodel_example
}
\description{
Includes all fitted models from vignette.
}
\keyword{datasets}
