% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOpenSCR.plot.R
\name{simOpenSCR.plot}
\alias{simOpenSCR.plot}
\title{Plot the open population simulation. Will document later.}
\usage{
simOpenSCR.plot(data, animated = FALSE, delay = 2)
}
\arguments{
\item{data}{an object created by simOpenSCR}
}
\value{
a plot
}
\description{
Plot the open population simulation. Will document later.
}
\author{
Ben Augustine
}
