% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cjs_openscr}
\alias{simulate_cjs_openscr}
\title{Simulate SCR open population Cormack-Jolly-Seber survey}
\usage{
simulate_cjs_openscr(par, N, n_occasions, detectors, mesh, move = FALSE,
  time = NULL, seed = NULL, print = TRUE)
}
\arguments{
\item{N}{number of individuals}

\item{n_occasions}{total number of occasions in survey}

\item{detectors}{secr trap object}

\item{mesh}{secr mesh object}

\item{move}{if TRUE then activity centres move and true_par$sd must be specified}

\item{time}{vector with time units between occasions}

\item{seed}{seed to set before simulating}

\item{print}{if TRUE then useful output is printed}

\item{true_par}{true parameters, named list of lambda0, sigma, phi}
}
\value{
ScrData object
}
\description{
Simulate SCR open population Cormack-Jolly-Seber survey
}
