% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_js_openscr}
\alias{simulate_js_openscr}
\title{Simulate SCR open population Jolly-Seber survey}
\usage{
simulate_js_openscr(par, n_occasions, detectors, mesh, move = FALSE,
  time = NULL, seed = NULL, print = TRUE)
}
\arguments{
\item{par}{true parameters, named list of lambda0, sigma, phi, beta}

\item{n_occasions}{total number of occasions in survey}

\item{detectors}{secr trap object}

\item{mesh}{secr mesh object}

\item{move}{if TRUE, activity centres move and par$sd must be specified}

\item{time}{vector with time units between occasions}

\item{seed}{seed to set before simulating}

\item{print}{if TRUE, useful output is printed}
}
\value{
ScrData object
}
\description{
Simulate SCR open population Jolly-Seber survey
}
