% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_scr}
\alias{simulate_scr}
\title{Simulate Spatial Capture-Recapture data}
\usage{
simulate_scr(par, n_occasions, detectors, mesh, move = FALSE,
  time = NULL, seed = NULL, print = TRUE)
}
\arguments{
\item{n_occasions}{number of occasions to simulate}

\item{detectors}{detectors object made using SECR package make.traps}

\item{mesh}{mesh object made using SECR make.mask}

\item{move}{if TRUE, then activity centres move by Brownian motion and par$sd must be specified}

\item{time}{time of each occasions (e.g. 1 to 10, or 2002, 2003, 2004)}

\item{seed}{if supplied, random number generator is given this seed}

\item{print}{if TRUE then useful output is printed}

\item{true_par}{named vector of D, lambda0, sigma parameters and sd, if movement is desired}
}
\value{
a ScrData object
}
\description{
Simulate Spatial Capture-Recapture data
}
