% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stoat}
\alias{stoat}
\title{Stoat data as included in secr package}
\format{A list of two data frames 
\describe{
  \item{captures}{Session, ID, Occasion, and Detector} 
  \item{traps}{Detector, x, and y}
}}
\usage{
stoat
}
\description{
This is a copy of the data included in the secr package. 
Data of A. E. Byrom from a study of stoats (Mustela erminea) in New Zealand. Individuals were identified from DNA in hair samples.
}
\keyword{datasets}
