% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_model.R
\docType{data}
\name{StrataModel}
\alias{StrataModel}
\title{Stratum model class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StrataModel
}
\description{
Fits the a model to different strata, sharing some parameters 
across strata. 
\itemize{
  \item scr_data: a list of ScrData objects, one per stratum  
  \item model: name of model class, e.g., ScrModel, CjsModel, JsModel
  \item shared_form: a named list of formulae for each parameter that is 
    shared across strata (~1 for constant)
  \item private_form: a named list of formula for each parameter that is private
  to each stratum, one per stratum
  \item print: (defualt TRUE) if TRUE then useful output is printed
  \item args : list of other arguments to passed to $new for model object 
}

Methods include: 
\itemize{
 \item get_par(...): calls get_par for each model object using arguments ... and
 returns list of results 
 \item calc_llk(): compute joint log-likelihood at current parameter values 
 \item fit: fit the model by obtaining the maximum likelihood estimates 
 \item par(): return current parameter of the model 
 \item mle_llk(): return log-likelihood value of maximum likelihood estimates
 \item is_shared(par): returns indices of any named elements in par that 
 are shared parameters across strata 
 \item is_private(par): returns indices of any named elements in par that are 
 private parameters across strata  
}
}
\keyword{datasets}
