% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratamodel_example}
\alias{stratamodel_example}
\title{Fitted models from Strata model vignette}
\format{A fitted model object  
\describe{
  \item{obj}{Strata model object}
}}
\usage{
stratamodel_example
}
\description{
Includes all fitted models from vignette.
}
\keyword{datasets}
