% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOpenSCR.R
\name{simOpenSCR}
\alias{simOpenSCR}
\title{Simulate data from a Open population SCR study.}
\usage{
simOpenSCR(N = c(40, 60, 80), gamma = NULL, phi = rep(0.8, 2),
  lam0 = rep(0.2, 3), sigma = rep(0.5, 3), K = rep(10, 3), X = X,
  t = 3, M = M, sigma_t = NULL, buff = 3, obstype = "bernoulli",
  ACtype = "fixed", vertices = NA, maxprop = 10000, primary = NA)
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate. If size 1, provide a gamma to determine N in subsequent years.
Otherwise, size is t, the number of years.}

\item{gamma}{a vector containing the per capita recruitment rates for each year.  If size 1, gamma is fixed across years.
Otherwise, size is t-1.  Do not enter a gamma if N for all years specified.}

\item{phi}{a vector containing the survival rates for each year. If size 1, phi is fixed across years.
Otherwise, size is t-1.}

\item{lam0}{a vector containing the detection function expected number of captures at distance 0. If size 1, constant rate is assumed.
Otherwise, size is t, the number of years.
Otherwise, size is t, the number of years.}

\item{sigma}{a vector containing the spatial scale parameter in each year.  If size 1, sigma is fixed across years.}

\item{K}{a vector containing the number of capture occasions in each year}

\item{X}{a list of trap locations in each year.  Each list element is a J[l] x 2 matrix of trap locations, with J[l] being
the number of traps in each year.}

\item{M}{an integer indicating the level of data augmentation to use during simulation.}

\item{sigma_t}{a numeric indicating the between year spatial scale parameter for ACtypes "metamu" "metamu2", and "markov"}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{obstype}{observation type, either "bernoulli" or "poisson"}

\item{ACtype}{Type of activity centers.  "fixed" don't move between years, "metamu" assume a bivariate normal distribution
with a meta mu and sigma_t with yearly activity centers required to stay inside the state space , "metamu2", is
the same as "metamu" except only meta mus are required to stay inside the state space.  markov" assumes activity
centers in year t+1 is a bivariate normal draw centered around the activity center in year t (but must stay within
the state space), and "independent" assumes animals randomly mix between years.}

\item{vertices}{a list of polygon vertices for each year if state space is not rectangular}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
This function simulates data from an open population SCR model.
}
\author{
Ben Augustine
}
