% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOpenSCRsex.R
\name{simOpenSCRsex}
\alias{simOpenSCRsex}
\title{Simulate data from a Open population SCR study with sex (or other group of size 2) differences.}
\usage{
simOpenSCRsex(N = cbind(c(30, 30), c(35, 45, 55)), pIDsex = 0.75,
  gamma = NULL, gamma.sex = TRUE, phi = rep(0.8, 2), lam0 = rep(0.2, 2),
  sigma = rep(0.5, 2), K = rep(10, 3), X = X, t = 3, M = M,
  sigma_t = NULL, buff = 3, obstype = "bernoulli", ACtype = "fixed",
  vertices = NA, maxprop = 10000)
}
\arguments{
\item{N}{a 1 x 2 vector or t x 2 matrix indicating the number of individuals to simulate. Column 1 is males, column 2 is females.
If only the first year is entered, provide a gamma to determine N in subsequent years.  Otherwise, size is t, the number of years.}

\item{phi}{a vector containing the survival rates for each year. If size 1, no sex difference is assumed.
Otherwise, sexes differ.}

\item{lam0}{a vector containing the detection function hazard rate in each year. If size 1, no sex difference is assumed.
Otherwise, sexes differ.}

\item{sigma}{a vector containing the spatial scale parameter in each year. If size 1, no sex difference is assumed.
Otherwise, sexes differ.  Do not enter a gamma if N for all years specified.}

\item{K}{a vector containing the number of capture occasions in each year}

\item{X}{a list of trap locations in each year.  Each list element is a J[l] x 2 matrix of trap locations, with J[l] being
the number of traps in each year.}

\item{M}{an integer indicating the level of data augmentation to use during simulation.}

\item{sigma_t}{a numeric indicating the between year spatial scale parameter for ACtypes "metamu" "metamu2", and "markov".  If size 1, no sex difference is assumed.
Otherwise, sexes differ.}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{obstype}{observation type, either "bernoulli" or "poisson"}

\item{ACtype}{Type of activity centers.  "fixed" don't move between years, "metamu" assume a bivariate normal distribution
with a meta mu and sigma_t with yearly activity centers required to stay inside the state space , "metamu2", is
the same as "metamu" except only meta mus are required to stay inside the state space.  markov" assumes activity
centers in year t+1 is a bivariate normal draw centered around the activity center in year t (but must stay within
the state space), and "independent" assumes animals randomly mix between years.}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
This function simulates data from an open population SCR model.
}
\author{
Ben Augustine
}
